organise pages and use Layout route for the header to appear on each page
