+++ components/Header.jsx
@@ -1,0 +1,17 @@
+import { Link } from "react-router-dom";
+
+const Header = () => {
+  return (
+    <header>
+      <Link className="site-logo" to="/">
+        #VanLife
+      </Link>
+      <nav>
+        <Link to="/about">About</Link>
+        <Link to="/vans">Vans</Link>
+      </nav>
+    </header>
+  );
+};
+
+export default Header;
