+++ components/Layout.jsx
@@ -1,0 +1,12 @@
+import Header from "./Header";
+import { Outlet } from "react-router-dom";
+const Layout = () => {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+};
+
+export default Layout;
