+++ src/App.jsx
@@ -1,27 +1,21 @@
-import { BrowserRouter, Routes, Route, Link } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "../pages/Home";
 import About from "../pages/About";
-import Vans from "../pages/Vans";
+import Vans from "../pages/Vans/Vans";
 import "../server"; // Import the server
-import VanDetail from "../pages/VanDetail";
+import VanDetail from "../pages/Vans/VanDetail";
+import Layout from "../components/Layout";
 const App = () => {
   return (
     <div>
       <BrowserRouter>
-        <header>
-          <Link className="site-logo" to="/">
-            #VanLife
-          </Link>
-          <nav>
-            <Link to="/about">About</Link>
-            <Link to="/vans">Vans</Link>
-          </nav>
-        </header>
         <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/about" element={<About />} />
-          <Route path="/vans" element={<Vans />} />
-          <Route path="/vans/:id" element={<VanDetail />} />
+          <Route element={<Layout />}>
+            <Route path="/" element={<Home />} />
+            <Route path="/about" element={<About />} />
+            <Route path="/vans" element={<Vans />} />
+            <Route path="/vans/:id" element={<VanDetail />} />
+          </Route>
         </Routes>
       </BrowserRouter>
     </div>
